0807 공부내용

+++ UnitTest/Character/Character_Demo.h
@@ -22,6 +22,6 @@ private:
 	float Fspeed;
 
 	float delay;
-	double angle_attack;
+	float angle_attack;
 
 };
